requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
dnslib==0.9.23
dnspython==2.3.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pycryptodome==3.18.0
PySocks==1.7.1
PyYAML==6.0
requests==2.31.0
scapy==2.5.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.2
webtech==1.3.3
